requirements.txt
gym==0.23.1
numpy==1.23.4
PyYAML==6.0
tensorboardX==2.5.1
torch==1.13.0
tqdm==4.64.1
pyyaml
